new base class utilized for HTTP targets
