+++ targets/Ciscosslvpn.py
@@ -1,9 +1,12 @@
 import csv
+from lib2to3.pytree import Base
 
 import requests
 
+from .classes.BaseHttpTarget import BaseHttpTarget
 
-class Ciscosslvpn:
+
+class Ciscosslvpn(BaseHttpTarget):
     def __init__(self, host, port, timeout, fireprox):
         self.group = input("Enter VPN group: ")
         self.timeout = timeout
@@ -68,41 +71,3 @@ class Ciscosslvpn:
             verify=False,
         )  # , proxies=self.proxyDict)
         return response
-
-    # handle CSV out output headers. Can be customized per module
-    def print_headers(self, csvfile):
-        # print table headers
-        print(
-            "%-35s %-17s %-13s %-15s"
-            % ("Username", "Password", "Response Code", "Response Length")
-        )
-        print("-" * 83)
-
-        # create CSV file
-        output = open(csvfile, "w")
-        fieldnames = ["Username", "Password", "Response Code", "Response Length"]
-        output_writer = csv.DictWriter(output, delimiter=",", fieldnames=fieldnames)
-        output_writer.writeheader()
-        output.close()
-
-    # handle target's response evaluation. Can be customized per module
-    def print_response(self, response, csvfile, timeout=False):
-        if timeout:
-            code = "TIMEOUT"
-            length = "TIMEOUT"
-        else:
-            code = response.status_code
-            length = str(len(response.content))
-
-        # print result to screen
-        print(
-            "%-35s %-17s %13s %15s"
-            % (self.data["username"], self.data["password"], code, length)
-        )
-
-        # print to CSV file
-        output = open(csvfile, "a")
-        output.write(
-            f'{self.data["username"]},{self.data["password"]},{code},{length}\n'
-        )
-        output.close()
