+++ targets/Citrix.py
@@ -1,11 +1,13 @@
 import requests
 
+from .classes.BaseHttpTarget import BaseHttpTarget
+
 ##############################
 # ALPHA CODE - NEEDS TESTING #
 ##############################
 
 
-class Citrix:
+class Citrix(BaseHttpTarget):
     def __init__(self, host, port, timeout, fireprox):
         self.timeout = timeout
         self.host = host
@@ -54,41 +56,3 @@ class Citrix:
             self.url, headers=self.headers, data=self.data, timeout=self.timeout
         )  # , verify=False, proxies=self.proxyDict)
         return response
-
-    # handle CSV out output headers. Can be customized per module
-    def print_headers(self, csvfile):
-        # print table headers
-        print(
-            "%-35s %-17s %-13s %-15s"
-            % ("Username", "Password", "Response Code", "Response Length")
-        )
-        print("-" * 83)
-
-        # create CSV file
-        output = open(csvfile, "w")
-        fieldnames = ["Username", "Password", "Response Code", "Response Length"]
-        output_writer = csv.DictWriter(output, delimiter=",", fieldnames=fieldnames)
-        output_writer.writeheader()
-        output.close()
-
-    # handle target's response evaluation. Can be customized per module
-    def print_response(self, response, csvfile, timeout=False):
-        if timeout:
-            code = "TIMEOUT"
-            length = "TIMEOUT"
-        else:
-            code = response.status_code
-            length = str(len(response.content))
-
-        # print result to screen
-        print(
-            "%-35s %-17s %13s %15s"
-            % (self.data["username"], self.data["password"], code, length)
-        )
-
-        # print to CSV file
-        output = open(csvfile, "a")
-        output.write(
-            f'{self.data["username"]},{self.data["password"]},{code},{length}\n'
-        )
-        output.close()
