+++ targets/Ntlm.py
@@ -3,11 +3,13 @@ import csv
 import requests
 from requests_ntlm import HttpNtlmAuth
 
+from .classes.BaseHttpTarget import BaseHttpTarget
 
-class Ntlm:
+
+class Ntlm(BaseHttpTarget):
     def __init__(self, host, port, timeout, path, fireprox):
         self.timeout = timeout
-        self.url = f"https://{host}:{port}/{path}"
+        self.url = f"http://{host}:{port}/{path}"
 
         if fireprox:
             self.url = f"https://{fireprox}/fireprox/{path}"
@@ -50,41 +52,3 @@ class Ntlm:
             self.url, headers=self.headers, auth=ntlm_auth, timeout=self.timeout
         )  # , verify=False, proxies=self.proxyDict)
         return response
-
-    # handle CSV out output headers. Can be customized per module
-    def print_headers(self, csvfile):
-        # print table headers
-        print(
-            "%-35s %-17s %-13s %-15s"
-            % ("Username", "Password", "Response Code", "Response Length")
-        )
-        print("-" * 83)
-
-        # create CSV file
-        output = open(csvfile, "w")
-        fieldnames = ["Username", "Password", "Response Code", "Response Length"]
-        output_writer = csv.DictWriter(output, delimiter=",", fieldnames=fieldnames)
-        output_writer.writeheader()
-        output.close()
-
-    # handle target's response evaluation. Can be customized per module
-    def print_response(self, response, csvfile, timeout=False):
-        if timeout:
-            code = "TIMEOUT"
-            length = "TIMEOUT"
-        else:
-            code = response.status_code
-            length = str(len(response.content))
-
-        # print result to screen
-        print(
-            "%-35s %-17s %13s %15s"
-            % (self.data["username"], self.data["password"], code, length)
-        )
-
-        # print to CSV file
-        output = open(csvfile, "a")
-        output.write(
-            f'{self.data["username"]},{self.data["password"]},{code},{length}\n'
-        )
-        output.close()
